backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
supabase>=0.7.1
python-dotenv>=0.19.0
aiohttp>=3.8.1
pyyaml>=6.0
watchdog>=2.1.9
psutil>=5.8.0
requests>=2.28.0
